bindgen: update kernel bindings to Linux 6.1
Switch to the ChromeOS v6.1 branch and regenerate all affected bindings
in the crosvm tree.

Some minor fixes were required in users of the bindings:
- KVM_SYSTEM_EVENT_RESET_FLAG_PSCI_RESET2 is available in the kernel now
  and was removed from the manually-added section of kvm_sys.
- The KVM system_event now contains an anonymous union, requiring a few
  renames.
- IORING_OP_* were moved into an enum, which required the enum name to
  be prefixed to the uses of those values.

BUG=None
TEST=tools/presubmit --all

Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4285742
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>

+++ hypervisor/src/kvm/aarch64.rs
@@ -220,7 +220,7 @@ impl KvmVcpu {
     /// `event_flags` should be one or more of the `KVM_SYSTEM_EVENT_RESET_FLAG_*` values defined by
     /// KVM.
     pub fn system_event_reset(&self, event_flags: u64) -> Result<VcpuExit> {
-        if event_flags & KVM_SYSTEM_EVENT_RESET_FLAG_PSCI_RESET2 != 0 {
+        if event_flags & u64::from(KVM_SYSTEM_EVENT_RESET_FLAG_PSCI_RESET2) != 0 {
             // Read reset_type and cookie from x1 and x2.
             let reset_type = self.get_one_reg(VcpuRegAArch64::X(1))?;
             let cookie = self.get_one_reg(VcpuRegAArch64::X(2))?;
